cambios

+++ routes/web.php
@@ -95,6 +95,7 @@ Route::get("/delete/duplicates", function(){
     $duplicates = DB::table('products') ->select('sku', DB::raw('COUNT(*) as `count`')) ->groupBy('sku') ->havingRaw('COUNT(*) > 1') ->get();
     foreach($duplicates as $duplicate){
         $product = App\Product::where("sku", $duplicate->sku)->first();
+        if($product)
         $product->delete();
     }
